requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.0.3
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==22.1.0
axial-positional-embedding==0.2.1
blessed==1.19.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.6
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.3.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.50.0
huggingface-hub==0.0.12
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==5.0.0
intel-openmp==2022.2.1
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl==2022.2.1
mkl-include==2022.2.1
multidict==6.0.2
networkx==2.8.8
ninja==1.11.1
nltk==3.7
numpy==1.23.4
nvidia-ml-py==11.495.46
oauthlib==3.2.2
omegaconf==2.1.0
opencv-python==4.5.3.56
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.1
pathtools==0.1.2
Pillow==8.3.1
promise==2.3
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-fast-transformers==0.4.0
pytorch-lightning==1.4.0
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-image==0.19.3
scipy==1.9.3
sentry-sdk==1.11.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
taming-transformers-rom1504==0.0.6
tbb==2021.7.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2022.10.10
tokenizers==0.10.3
torch==1.7.1+cu110
torchaudio==0.7.2
torchmetrics==0.4.1
torchvision==0.8.2+cu110
torchxrayvision==0.0.39
tqdm==4.64.1
transformers==4.9.1
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.5
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
